chore: rename directory to profiles_mlcorelib (#324)
