+++ README.md
@@ -72,4 +72,4 @@ py_models:
 
 ### 3. Building the model
 
-Refer to our [docs page](https://www.rudderstack.com/docs/profiles/predictions/) for how to set up a python model in your project. There are various advanced config options that you can find in the [model_configs.yaml](src/predictions/rudderstack_predictions/config/model_configs.yaml) file. You can add these options in the python model of your profiles project to override the defaults.
+Refer to our [docs page](https://www.rudderstack.com/docs/profiles/predictions/) for how to set up a python model in your project. There are various advanced config options that you can find in the [model_configs.yaml](src/predictions/profiles_mlcorelib/config/model_configs.yaml) file. You can add these options in the python model of your profiles project to override the defaults.
