+++ tests/integration/bigquery/regressor.py
@@ -3,8 +3,8 @@ from train import *
 import shutil
 import time
 from predict import *
-from src.predictions.rudderstack_predictions.wht.rudderPB import RudderPB
-from src.predictions.rudderstack_predictions.connectors.ConnectorFactory import (
+from src.predictions.profiles_mlcorelib.wht.rudderPB import RudderPB
+from src.predictions.profiles_mlcorelib.connectors.ConnectorFactory import (
     ConnectorFactory,
 )
 import json
