+++ tests/integration/redshift/regressor.py
@@ -2,7 +2,7 @@ from train import *
 import shutil
 import time
 from predict import *
-from src.predictions.rudderstack_predictions.connectors.RedshiftConnector import (
+from src.predictions.profiles_mlcorelib.connectors.RedshiftConnector import (
     RedshiftConnector,
 )
 import json
