+++ tests/unit/SnowflakeConnector.py
@@ -1,5 +1,5 @@
 # Generated by CodiumAI
-from src.predictions.rudderstack_predictions.connectors.SnowflakeConnector import (
+from src.predictions.profiles_mlcorelib.connectors.SnowflakeConnector import (
     SnowflakeConnector,
 )
 import pandas as pd
@@ -126,11 +126,11 @@ class TestValidations(unittest.TestCase):
 
     # Checks if no:of columns in the feature table is less than 3, then it raises an exception.
     @patch(
-        "src.predictions.rudderstack_predictions.utils.constants.CLASSIFIER_MIN_LABEL_PROPORTION",
+        "src.predictions.profiles_mlcorelib.utils.constants.CLASSIFIER_MIN_LABEL_PROPORTION",
         new=1.0,
     )
     @patch(
-        "src.predictions.rudderstack_predictions.utils.constants.CLASSIFIER_MAX_LABEL_PROPORTION",
+        "src.predictions.profiles_mlcorelib.utils.constants.CLASSIFIER_MAX_LABEL_PROPORTION",
         new=0.0,
     )
     def test_expects_error_if_label_ratios_are_bad_classification(self):
@@ -160,11 +160,11 @@ class TestValidations(unittest.TestCase):
             )
 
     @patch(
-        "src.predictions.rudderstack_predictions.utils.constants.CLASSIFIER_MIN_LABEL_PROPORTION",
+        "src.predictions.profiles_mlcorelib.utils.constants.CLASSIFIER_MIN_LABEL_PROPORTION",
         new=0.05,
     )
     @patch(
-        "src.predictions.rudderstack_predictions.utils.constants.CLASSIFIER_MAX_LABEL_PROPORTION",
+        "src.predictions.profiles_mlcorelib.utils.constants.CLASSIFIER_MAX_LABEL_PROPORTION",
         new=0.95,
     )
     def test_passes_for_good_data_classification(self):
@@ -181,7 +181,7 @@ class TestValidations(unittest.TestCase):
         self.assertTrue(self.connector.validate_class_proportions(table, "COL1"))
 
     @patch(
-        "src.predictions.rudderstack_predictions.utils.constants.REGRESSOR_MIN_LABEL_DISTINCT_VALUES",
+        "src.predictions.profiles_mlcorelib.utils.constants.REGRESSOR_MIN_LABEL_DISTINCT_VALUES",
         new=3,
     )
     def test_passes_for_good_data_regression(self):
