+++ tests/unit/constants.py
@@ -1,7 +1,7 @@
 # Generated by CodiumAI
 
 import unittest
-from src.predictions.rudderstack_predictions.utils.constants import TrainTablesInfo
+from src.predictions.profiles_mlcorelib.utils.constants import TrainTablesInfo
 
 
 class TestTrainTablesInfo(unittest.TestCase):
