+++ tests/unit/pythonWHT.py
@@ -2,14 +2,14 @@ import unittest
 
 from datetime import datetime
 from unittest.mock import patch, Mock
-from src.predictions.rudderstack_predictions.wht.pythonWHT import PythonWHT
+from src.predictions.profiles_mlcorelib.wht.pythonWHT import PythonWHT
 
 
 class TestGetInputModels(unittest.TestCase):
     def setUp(self) -> None:
         self.pythonWHT = PythonWHT()
 
-    @patch("src.predictions.rudderstack_predictions.wht.rudderPB.RudderPB.show_models")
+    @patch("src.predictions.profiles_mlcorelib.wht.rudderPB.RudderPB.show_models")
     def test_get_input_models(self, mock_rudderpb_show_models):
         original_input_models = [
             "Material_user_var_table1_54ddc22a_383",
@@ -77,7 +77,7 @@ class TestFetchValidHistoricMaterials(unittest.TestCase):
         ]
 
     @patch(
-        "src.predictions.rudderstack_predictions.wht.pythonWHT.PythonWHT._validate_historical_materials_hash"
+        "src.predictions.profiles_mlcorelib.wht.pythonWHT.PythonWHT._validate_historical_materials_hash"
     )
     def test_all_data_present_and_valid(self, mock_validate_historical_materials_hash):
         # Mock dependencies
@@ -107,7 +107,7 @@ class TestFetchValidHistoricMaterials(unittest.TestCase):
         self.assertEqual(len(materials), 1)
 
     @patch(
-        "src.predictions.rudderstack_predictions.wht.pythonWHT.PythonWHT._validate_historical_materials_hash"
+        "src.predictions.profiles_mlcorelib.wht.pythonWHT.PythonWHT._validate_historical_materials_hash"
     )
     def test_missing_sequence_number(self, mock_compute_material_name):
         connector = MockConnector()
